CORE-3663 Fixed warning about failing to cd home
This applies to all backend services.  The runuser command was
attempting to log in as a given user.  This caused an attempt to change
the current directory to the user's home directory.  The iplant system
user doesn't have a home directory, thus the warning.  Removing the -l
option fixes this.